[helm] Update ArgoCD configurations to use 'default' namespace for production
This commit modifies the ArgoCD configuration files to set the namespace to 'default' for production environments instead of 'splattop-prod'. Changes include updates to the README.md, splattop-applicationset.yaml, splattop-prod.yaml, and splattop-project.yaml files. This adjustment streamlines the deployment process by aligning with existing production workloads and enhances clarity in the configuration documentation.